backend/requirements.txt
fastapi
uvicorn
websockets
requests
python-multipart
openai
elevenlabs
numpy
groq
python-dotenv
deepgram-sdk
spotipy
yt-dlp